+++ src/execution/index/art/prefix.cpp
@@ -67,110 +67,6 @@ void Prefix::New(ART &art, reference<NodePtr> &node_ref, const ARTKey &key, cons
 	}
 }
 
-PrefixHandle PrefixHandle::NewInternal(ART &art, NodePtr &node, const_data_ptr_t data, const uint8_t count,
-                                       const idx_t offset) {
-	node = NodePtr::GetAllocator(art, PREFIX).New();
-	node.SetMetadata(static_cast<uint8_t>(PREFIX));
-
-	PrefixHandle prefix(NodeHandle(art, node));
-	prefix.SetCount(art, count);
-	if (data) {
-		D_ASSERT(count);
-		memcpy(prefix.Data(), data + offset, count);
-	}
-	prefix.Child(art).Clear();
-	return prefix;
-}
-
-PrefixChain PrefixHandle::New(ART &art, const ARTKey &key, const idx_t depth, const idx_t count) {
-	D_ASSERT(count > 0);
-
-	NodePtr root;
-	auto first_count = UnsafeNumericCast<uint8_t>(MinValue<idx_t>(art.PrefixCount(), count));
-	auto prefix = NewInternal(art, root, key.data, first_count, depth);
-	auto tail = std::move(prefix).IntoChild(art);
-
-	idx_t offset = first_count;
-	while (offset < count) {
-		auto this_count = UnsafeNumericCast<uint8_t>(MinValue<idx_t>(art.PrefixCount(), count - offset));
-		auto next = NewInternal(art, tail.Get(), key.data, this_count, depth + offset);
-		tail = std::move(next).IntoChild(art);
-
-		offset += this_count;
-	}
-	return {root, std::move(tail)};
-}
-
-PrefixHandle PrefixHandle::AppendByte(ART &art, PrefixHandle prefix, const uint8_t byte) {
-	const auto count = prefix.GetCount(art);
-	if (count != art.PrefixCount()) {
-		prefix.SetByte(count, byte);
-		prefix.SetCount(art, UnsafeNumericCast<uint8_t>(count + 1));
-		return prefix;
-	}
-
-	auto tail = std::move(prefix).IntoChild(art);
-	return NewInternal(art, tail.Get(), &byte, 1, 0);
-}
-
-void PrefixHandle::Append(ART &art, PrefixHandle prefix, NodePtr other) {
-	D_ASSERT(other.HasMetadata());
-
-	while (other.GetType() == PREFIX) {
-		if (other.GetGateStatus() == GateStatus::GATE_SET) {
-			prefix.Child(art) = other;
-			return;
-		}
-
-		NodePtr next;
-		{
-			PrefixHandle other_prefix(NodeHandle(art, other));
-			const auto count = other_prefix.GetCount(art);
-			for (idx_t i = 0; i < count; i++) {
-				prefix = AppendByte(art, std::move(prefix), other_prefix.GetByte(i));
-			}
-			next = other_prefix.Child(art);
-			prefix.Child(art) = next;
-		}
-
-		NodePtr::FreeNode(art, other);
-		other = next;
-	}
-	prefix.Child(art) = other;
-}
-
-NodePtr PrefixHandle::Split(ART &art, NodePtr &node, NodePtr &replacement, const uint8_t pos) {
-	D_ASSERT(node.HasMetadata());
-	D_ASSERT(node.GetType() == PREFIX);
-	D_ASSERT(replacement.HasMetadata());
-
-	NodePtr child;
-	{
-		PrefixHandle prefix(NodeHandle(art, node));
-		const auto count = prefix.GetCount(art);
-		D_ASSERT(pos < count);
-
-		if (pos + 1 < count) {
-			const auto suffix_count = UnsafeNumericCast<uint8_t>(count - pos - 1);
-			auto suffix = NewInternal(art, child, prefix.Data(), suffix_count, pos + 1);
-			Append(art, std::move(suffix), prefix.Child(art));
-		} else {
-			child = prefix.Child(art);
-		}
-
-		if (pos != 0) {
-			prefix.SetCount(art, pos);
-			prefix.Child(art) = replacement;
-			return child;
-		}
-		replacement.SetGateStatus(node.GetGateStatus());
-	}
-
-	NodePtr::FreeNode(art, node);
-	node = replacement;
-	return child;
-}
-
 void Prefix::Concat(ART &art, NodePtr &parent, NodePtr &node4, const NodePtr child, uint8_t byte,
                     const GateStatus node4_status, const GateStatus status) {
 	// We have four situations from which we enter here:
